Fix User type test in MyRadioEmail.
The previous code tested against a class that no longer exists, and
also used 'is_a', which seems to be discouraged in favour of
'instanceof'.

This commit changes the class to the correct class, MyRadio_User,
and also changes from is_a to instanceof.  The exception has also
been slightly reworded.